video_intelligence_mvp/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
opencv-python-headless
python-dotenv
python-multipart
